check before deleting base path

+++ tests/test_rpc_cli.rs
@@ -24,7 +24,10 @@ const KEY_STORE_PATH: &str = "./tmp/test_rpc_cli/key_store";
 fn test_service_ready() -> bool {
     let mut base_path = Path::new(TMP_DIR).to_owned();
     base_path.push("base");
-    std::fs::remove_dir_all(base_path.clone()).unwrap();
+    if base_path.exists() {
+        std::fs::remove_dir_all(base_path.clone()).unwrap();
+    }
+
     thread::spawn(move || { devnet::start_devnet(Some(&base_path)) });
     thread::sleep(Duration::from_secs(1));
     true
